// web/src/api/generated/model/updateRun.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Firetower
 * The Firetower control plane: API, scheduling, and worker transports.
 * OpenAPI spec version: 0.30.1
 */
import type { RunState } from './runState';
import type { Targets } from './targets';
import type { UpdateStep } from './updateStep';

export interface UpdateRun {
  createdAt: string;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  fromVersion: string;
  id: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  startedBy?: string | null;
  state: RunState;
  steps: UpdateStep[];
  targets: Targets;
  toVersion: string;
  whenIdle: boolean;
}