// web/src/api/generated/model/updateStep.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Firetower
 * The Firetower control plane: API, scheduling, and worker transports.
 * OpenAPI spec version: 0.32.2
 */
import type { StepState } from './stepState';

export interface UpdateStep {
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  log: string;
  position: number;
  /** @nullable */
  startedAt?: string | null;
  state: StepState;
  target: string;
  title: string;
}